R script to calculate burst coordination
R script used to quantify the coordination in timing between bursts in different channels of the same file. Input is "Collated_data.xlsx" and output is another spreadsheet named "burst_coord_data.xlsx", which includes injury, drug condition, file number, and the percent of time that bursts co-occur (calculated as the number of samples shared between bursts in separate channels over the total number of samples that include bursts).